refactor(trusty-common): split embedder_client/supervisor.rs under the 500-SLOC cap
CI flagged supervisor.rs at 510 SLOC (>500 prod cap, not allowlisted).
Extract the self-contained SupervisorConfig/batch-size-tuning section
(SupervisorConfig, cuda_sidecar_batch_cap, sidecar_batch_size, parse_env)
into a new supervisor_config.rs sibling module — pure config plumbing with
no dependency on the child-process/watch-channel supervision machinery.
Re-exported from supervisor.rs so embedder_client's public API
(SupervisorConfig, cuda_sidecar_batch_cap, sidecar_batch_size) and
supervisor_tests.rs's `use super::*;` are unaffected.

Epic #3524 slice 5 CI follow-up.
